The gone is gone.
Don't you hate it when you are walking along a long plank and all the sudden poof! You can't see the plank any more because the world decided to unload it because the points of the plank are technically out of view? Well hate no more! I implemented a fix for that! yay
New:
-Set outline color of a quobject
-Added zNear clipping
-getPlane method in Math3D
-randomColors in quobjects will set a random color for every face
-New staff.obj that I made poorly in blender
Fixed:
-Quobjects disappearing if too big.
-Some improvement to quaphics class
Next to come:
-UV mapping
-Light source
-Better lighting